requirements/common.txt
Django==2.0.1
djangorestframework==3.7.7
drf-extensions==0.3.1
envparse==0.2.0
boto3==1.5.12
django-storages==1.6.5
django_redis==4.8.0
psycopg2==2.7.3.2
raven==6.5.0
newrelic==2.100.0.84
